moved the tests under the correct package #243
